Moved product related and site related methods into their own class from `WCOrderTestUtils`
